Added Lambda Functionality
This method receives a POST requests and converts a long URL into a minified form. To do this, the contents are hashed using Python's standard hashing method and then, the numeric result is encoded in base62. Finally, this result is put into a DynamoDB table as a pair of [minified : full url]